finance-agent/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
scikit-learn
httpx
